just some touch ups for the demo
I added a BUGS file where we can put bugs that we find that can act as a scrum board, oh noo OOP jargon!.  Added a toggle switch for the flashlight mostly for me so i can check out emissive lights in the future a little better and fixed a material problem on the chair. and disabled comments. maybe other stuff but this is a long description