Final Push for now
The lighting is much better now with GI and the the fire is much better as well. the colors of the fire still need adjusted to be more friendly if that makes sense (ill probably just make a new fire) and I need to make the god rays come back, probably as an artificial god ray. The artificial ray will be lest computationally expensive too!