Start of game updated
You no longer start with flashlight but rather have to pick it up off ground in front of you. Particle system to indicate when looking at it to give feedback letting the player know that you can interact with it. Picking it up puts it in your hand and removes it form the ground